refactor: simplify top-level tetra modules (ResourceManager + ServerlessResource)
