+++ tetra/__init__.py
@@ -2,8 +2,8 @@
 
 from .client import RemoteExecutionClient, remote
 from . import remote_execution_pb2, remote_execution_pb2_grpc
-from .resource_manager import ResourceManager
-
+from .core.resources.resource_manager import ResourceManager
+from .core.resources.serverless import ServerlessResource
 
 __all__ = [
     "RemoteExecutionClient",
@@ -11,4 +11,5 @@ __all__ = [
     "remote_execution_pb2",
     "remote_execution_pb2_grpc",
     "ResourceManager",
+    "ServerlessResource",
 ]
